cv-engine/test_corrected_system.py
#!/usr/bin/env python3
"""
Test the corrected face recognition system
"""

import requests
import json
import os
from PIL import Image
import io

def test_face_recognition_api():
    """Test the face recognition endpoints"""
    base_url = "http://127.0.0.1:5001"
    
    print("🧪 Testing Corrected Face Recognition System")
    print("=" * 60)
    
    # Test 1: Debug embeddings
    print("\n1️⃣ Testing debug embeddings endpoint:")
    try:
        response = requests.get(f"{base_url}/debug-embeddings")
        if response.status_code == 200:
            data = response.json()
            print(f"✅ Current enrolled students: {data['student_ids']}")
            print(f"✅ Total encodings: {data['encodings_count']}")
        else:
            print(f"❌ Error: {response.status_code} - {response.text}")
    except Exception as e:
        print(f"❌ Connection error: {e}")
    
    # Test 2: Face verification with each student's image
    print(f"\n2️⃣ Testing face verification:")
    
    students = ['132', '172', '564']
    uploads_dir = 'uploads/students'
    
    for student_id in students:
        print(f"\n   Testing student {student_id}:")
        image_path = os.path.join(uploads_dir, f'{student_id}.jpg')
        
        if os.path.exists(image_path):
            try:
                with open(image_path, 'rb') as f:
                    files = {'photo': f}
                    response = requests.post(f"{base_url}/verify", files=files)
                
                if response.status_code == 200:
                    data = response.json()
                    if data['success']:
                        recognized = data['recognized']
                        print(f"     ✅ Recognized: {len(recognized)} students")
                        for rec in recognized:
                            print(f"        • Student {rec['student_id']}: {rec['confidence']:.3f} confidence")
                    else:
                        print(f"     ❌ No recognition: {data['message']}")
                else:
                    print(f"     ❌ Error: {response.status_code} - {response.text}")
                    
            except Exception as e:
                print(f"     ❌ Error testing {student_id}: {e}")
        else:
            print(f"     ❌ Image not found: {image_path}")
    
    print(f"\n💡 Expected Results:")
    print("• Each student should be recognized as themselves")
    print("• No false positive recognitions")
    print("• Confidence scores should be reasonable (>0.5)")

def test_similarity_with_new_thresholds():
    """Test the new similarity thresholds manually"""
    print(f"\n3️⃣ Testing new similarity thresholds:")
    
    import cv2
    import numpy as np
    import pickle
    
    # Load current student data
    with open('student_embeddings.pkl', 'rb') as f:
        student_data = pickle.load(f)
    
    def compare_opencv_encodings(enc1, enc2, tolerance=0.6):
        """Same comparison as server"""
        enc1_norm = enc1 / (np.linalg.norm(enc1) + 1e-7)
        enc2_norm = enc2 / (np.linalg.norm(enc2) + 1e-7)
        
        similarity = np.dot(enc1_norm, enc2_norm)
        
        # Use same logic as server
        if tolerance <= 0.4:  # For duplicate detection
            threshold = 0.99  # 99% similarity required
        else:  # For recognition
            threshold = 0.75  # 75% similarity required
        
        print(f"     Similarity: {similarity:.4f}, Threshold: {threshold}, Match: {similarity > threshold}")
        return similarity > threshold
    
    students = student_data['student_ids']
    encodings = student_data['encodings']
    
    print(f"   Enrolled students: {students}")
    
    # Test duplicate detection threshold (tolerance=0.4)
    print(f"\n   🔍 Duplicate Detection Tests (99% threshold):")
    for i, student1 in enumerate(students):
        for j, student2 in enumerate(students):
            if i < j:
                print(f"     {student1} vs {student2}:")
                compare_opencv_encodings(encodings[i], encodings[j], tolerance=0.4)
    
    # Test recognition threshold (tolerance=0.6)  
    print(f"\n   🎯 Recognition Tests (75% threshold):")
    for i, student1 in enumerate(students):
        for j, student2 in enumerate(students):
            if i < j:
                print(f"     {student1} vs {student2}:")
                compare_opencv_encodings(encodings[i], encodings[j], tolerance=0.6)

if __name__ == "__main__":
    test_face_recognition_api()
    test_similarity_with_new_thresholds()
    
    print(f"\n📋 Summary:")
    print("✅ System reset completed with manual override")
    print("✅ Ultra-strict thresholds implemented (99% for duplicates)")
    print("✅ All three students enrolled as separate individuals")
    print("⚠️  OpenCV method still has limitations - proper face_recognition library recommended")