cv-engine/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0